# docs/swagger.yaml
basePath: /
definitions:
  category.Category:
    properties:
      id:
        description: Category ID
        type: string
      level:
        description: Depth level
        type: integer
      nameen:
        description: Name English
        type: string
      nameru:
        description: Name Russian
        type: string
      nametm:
        description: Name Turkmen
        type: string
      nametr:
        description: Name Turkish
        type: string
      organization:
        description: Organization ID
        type: string
      parent:
        description: Parent category ID
        type: string
    required:
    - nameen
    - nameru
    - nametm
    - nametr
    - organization
    type: object
  category.CreateCategoryInput:
    properties:
      level:
        description: Depth level
        type: integer
      nameen:
        description: Name English
        type: string
      nameru:
        description: Name Russian
        type: string
      nametm:
        description: Name Turkmen
        type: string
      nametr:
        description: Name Turkish
        type: string
      organization:
        description: Organization ID
        type: string
      parent:
        description: Parent category ID
        type: string
    required:
    - nameen
    - nameru
    - nametm
    - nametr
    - organization
    type: object
  category.UpdateCategoryInput:
    properties:
      id:
        description: Category ID
        type: string
      level:
        description: Depth level
        type: integer
      nameen:
        description: Name English
        type: string
      nameru:
        description: Name Russian
        type: string
      nametm:
        description: Name Turkmen
        type: string
      nametr:
        description: Name Turkish
        type: string
      organisation:
        description: Organization ID
        type: string
      parent:
        description: Parent category ID
        type: string
    type: object
...
  http.AuthResponse:
    properties:
      tokens:
        $ref: '#/definitions/token.Tokens'
        description: Tokens
      user:
        $ref: '#/definitions/user.User'
        description: User
    type: object
  http.ErrorResponse:
    properties:
      message:
        description: Error message
        type: string
    type: object
  http.StatusResponse:
    properties:
      status:
        description: Status message
        type: string
    type: object
...
  token.RefreshToken:
    properties:
      token:
        description: Refresh Token
        type: string
    type: object
  token.Tokens:
    properties:
      accessToken:
        description: Access Token
        type: string
      accessTokenExpires:
        description: Access Token Expires datetime
        type: integer
      refreshToken:
        description: Refresh Token
        type: string
      tokenType:
        description: Token Type
        type: string
    type: object
  user.CreateUserInput:
    properties:
      firstname:
        description: First name
        type: string
      lastname:
        description: Last Name
        type: string
      password:
        description: Password
        type: string
      phone:
        description: Phone number
        type: string
      roleid:
        description: Role ID
        type: integer
    required:
    - firstname
    - lastname
    - password
    - phone
    type: object
  user.SignInInput:
    properties:
      password:
        description: Users password
        type: string
      phone:
        description: Users phone number
        type: string
    required:
    - password
    - phone
    type: object
...
  user.User:
    properties:
      firstname:
        description: First name
        type: string
      id:
        description: User ID
        type: string
      lastname:
        description: Last Name
        type: string
      password:
        description: Password
        type: string
      phone:
        description: Phone number
        type: string
      role:
        description: Role name
        type: string
      roleid:
        description: Role ID
        type: integer
      status:
        description: Users status
        type: string
    required:
    - firstname
    - lastname
    - password
    - phone
    type: object
info:
  contact:
    email: evgeniydammer@gmail.com
    name: Evgeniy Dammer
  description: marketplace API service
  title: marketplace API
  version: "1.0"
paths:
  /api/v1/categories/:
    patch:
      consumes:
      - application/json
      description: Update category method.
      parameters:
      - description: Category data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/category.UpdateCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Update category method.
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create category method.
      parameters:
      - description: Category data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Category ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Create category method.
      tags:
      - categories
  /api/v1/categories/{org_id}:
    get:
      consumes:
      - application/json
      description: Get all categories method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category List
          schema:
            items:
              $ref: '#/definitions/category.Category'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all categories method.
      tags:
      - categories
  /api/v1/categories/{org_id}/{id}:
    delete:
      consumes:
      - application/json
      description: Delete category method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete category method.
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get category by id method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category data
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get category by id method.
      tags:
      - categories
...
  /refresh/:
    post:
      consumes:
      - application/json
      description: Refresh token method.
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/token.RefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: User data and Tokens
          schema:
            $ref: '#/definitions/http.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Refresh token method.
      tags:
      - authentication
  /signin/:
    post:
      consumes:
      - application/json
      description: SignIn user method.
      parameters:
      - description: Username and Password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: User data and Tokens
          schema:
            $ref: '#/definitions/http.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: SignIn user method.
      tags:
      - authentication
  /signup/:
    post:
      consumes:
      - application/json
      description: SignUp user method.
      parameters:
      - description: User data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: User ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: SignUp user method.
      tags:
      - authentication
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"